Move ItemPageTable.js to /item-pages/components/
