+++ src/encoded/static/components/browse/components/index.js
@@ -10,7 +10,6 @@ export { SortController } from './SortController';
 export { SelectedFilesController } from './SelectedFilesController';
 export { CustomColumnController, CustomColumnSelector } from './CustomColumnController';
 export { SearchResultTable } from './SearchResultTable';
-export { ItemPageTable, ItemPageTableLoader, ItemPageTableBatchLoader, ItemPageTableSearchLoaderPageController, ItemPageTableSearchLoader } from './ItemPageTable';
 export { AboveTableControls } from './AboveTableControls';
 export { AboveSearchTablePanel } from './AboveSearchTablePanel';
 export { ExperimentSetDetailPane } from './ExperimentSetDetailPane';
