+++ src/encoded/static/components/item-pages/BiosampleView.js
@@ -5,10 +5,10 @@ import PropTypes from 'prop-types';
 import _ from 'underscore';
 import * as globals from './../globals';
 import { console, object, expFxn, ajax, Schemas, layout, fileUtil, isServerSide, DateUtility } from './../util';
-import { FormattedInfoBlock, ExperimentSetTablesLoadedFromSearch } from './components';
+import { FormattedInfoBlock, ExperimentSetTablesLoadedFromSearch, ItemPageTable } from './components';
 import DefaultItemView, { OverViewBodyItem } from './DefaultItemView';
 import { IndividualItemTitle } from './BiosourceView';
-import { ExperimentSetDetailPane, ResultRowColumnBlockValue, ItemPageTable } from './../browse/components';
+import { ExperimentSetDetailPane, ResultRowColumnBlockValue } from './../browse/components';
 
 
 export default class BiosampleView extends DefaultItemView {
