+++ src/encoded/static/components/item-pages/ExperimentView.js
@@ -9,7 +9,7 @@ import { console, object, expFxn, ajax, Schemas, layout, fileUtil, isServerSide 
 import { FormattedInfoBlock, TabbedView, ExperimentSetTables, ExperimentSetTablesLoaded, WorkflowNodeElement,
     SimpleFilesTableLoaded, SimpleFilesTable, Publications, OverviewHeadingContainer } from './components';
 import { OverViewBodyItem } from './DefaultItemView';
-import { ExperimentSetDetailPane, ResultRowColumnBlockValue, ItemPageTable } from './../browse/components';
+import { ExperimentSetDetailPane, ResultRowColumnBlockValue } from './../browse/components';
 import Graph, { parseAnalysisSteps, parseBasicIOAnalysisSteps } from './../viz/Workflow';
 import { requestAnimationFrame } from './../viz/utilities';
 import { RowSpacingTypeDropdown } from './WorkflowView';
