+++ src/encoded/static/components/item-pages/components/index.js
@@ -30,3 +30,4 @@ export { SimpleFilesTable, SimpleFilesTableLoaded } from './SimpleFilesTable';
 export { AdjustableDividerRow, DraggableVerticalBorder } from './AdjustableDividerRow';
 export { CollapsibleItemViewButtonToolbar } from './CollapsibleItemViewButtonToolbar';
 export { OverviewHeadingContainer } from './OverviewHeadingContainer';
+export { ItemPageTable, ItemPageTableLoader, ItemPageTableBatchLoader, ItemPageTableSearchLoaderPageController, ItemPageTableSearchLoader } from './ItemPageTable';
